+++ engine/runtime/src/main/java/org/enso/interpreter/node/callable/InvokeConversionNode.java
@@ -172,10 +172,10 @@ public abstract class InvokeConversionNode extends BaseNode {
       }
     }
     arguments[thatArgumentPosition] = that.getValue();
-    ArrayRope<Warning> warnings = that.getReassignedWarnings(this);
+    ArrayRope<Warning> warnings = that.getReassignedWarningsAsRope(this);
     Object result =
         childDispatch.execute(frame, state, conversion, self, that.getValue(), arguments);
-    return WithWarnings.prependTo(result, warnings);
+    return WithWarnings.appendTo(result, warnings);
   }
 
   @Specialization(guards = "interop.isString(that)")