+++ engine/runtime/src/main/java/org/enso/interpreter/runtime/error/WithWarnings.java
@@ -1,5 +1,6 @@
 package org.enso.interpreter.runtime.error;
 
+import com.oracle.truffle.api.CompilerDirectives;
 import org.enso.interpreter.runtime.data.ArrayRope;
 import org.enso.interpreter.runtime.library.dispatch.TypesLibrary;
 
@@ -11,22 +12,27 @@ import com.oracle.truffle.api.library.ExportMessage;
 import com.oracle.truffle.api.library.Message;
 import com.oracle.truffle.api.library.ReflectionLibrary;
 import com.oracle.truffle.api.nodes.Node;
+import org.graalvm.collections.EconomicSet;
+import org.graalvm.collections.Equivalence;
+
+import java.util.Arrays;
 
 @ExportLibrary(TypesLibrary.class)
 @ExportLibrary(WarningsLibrary.class)
 @ExportLibrary(ReflectionLibrary.class)
 public final class WithWarnings implements TruffleObject {
-  private final ArrayRope<Warning> warnings;
+  private final EconomicSet<Warning> warnings;
   private final Object value;
 
   private WithWarnings(Object value, Warning... warnings) {
     assert !(value instanceof WithWarnings);
-    this.warnings = new ArrayRope<>(warnings);
+    this.warnings = createSetFromArray(warnings);
     this.value = value;
   }
 
-  private WithWarnings(Object value, ArrayRope<Warning> warnings) {
-    this.warnings = warnings;
+  private WithWarnings(Object value, EconomicSet<Warning> warnings, Warning... additionalWarnings) {
+    assert !(value instanceof WithWarnings);
+    this.warnings = cloneSetAndAppend(warnings, additionalWarnings);
     this.value = value;
   }
 
@@ -43,36 +49,25 @@ public final class WithWarnings implements TruffleObject {
   }
 
   public WithWarnings append(Warning... newWarnings) {
-    return new WithWarnings(value, warnings.append(newWarnings));
+    return new WithWarnings(value, warnings, newWarnings);
   }
 
   public WithWarnings append(ArrayRope<Warning> newWarnings) {
-    return new WithWarnings(value, warnings.append(newWarnings));
-  }
-
-  public WithWarnings prepend(Warning... newWarnings) {
-    return new WithWarnings(value, warnings.prepend(newWarnings));
-  }
-
-  public WithWarnings prepend(ArrayRope<Warning> newWarnings) {
-    return new WithWarnings(value, warnings.prepend(newWarnings));
+    return new WithWarnings(value, warnings, newWarnings.toArray(Warning[]::new));
   }
 
   public Warning[] getWarningsArray(WarningsLibrary warningsLibrary) {
-    Warning[] warningsArr = warnings.toArray(Warning[]::new);
     Warning[] allWarnings;
-
     if (warningsLibrary != null && warningsLibrary.hasWarnings(value)) {
       try {
-        Warning[] valuesWarnings = warningsLibrary.getWarnings(value, null);
-        allWarnings = new Warning[valuesWarnings.length + warningsArr.length];
-        System.arraycopy(warningsArr, 0, allWarnings, 0, warningsArr.length);
-        System.arraycopy(valuesWarnings, 0, allWarnings, warningsArr.length, valuesWarnings.length);
+        Warning[] valueWarnings = warningsLibrary.getWarnings(value, null);
+        EconomicSet<Warning> tmp = cloneSetAndAppend(warnings, valueWarnings);
+        allWarnings = Warning.fromSetToArray(tmp);
       } catch (UnsupportedMessageException e) {
         throw new IllegalStateException(e);
       }
     } else {
-      allWarnings = warningsArr;
+      allWarnings = Warning.fromSetToArray(warnings);
     }
     return allWarnings;
   }
@@ -82,37 +77,29 @@ public final class WithWarnings implements TruffleObject {
     return warnings.size();
   }
 
-  public ArrayRope<Warning> getReassignedWarnings(Node location) {
-    return getReassignedWarnings(location, null);
+  public ArrayRope<Warning> getReassignedWarningsAsRope(Node location) {
+    return new ArrayRope<>(getReassignedWarnings(location, null));
   }
 
-  public ArrayRope<Warning> getReassignedWarnings(Node location, WarningsLibrary warningsLibrary) {
+  public Warning[] getReassignedWarnings(Node location, WarningsLibrary warningsLibrary) {
     Warning[] warnings = getWarningsArray(warningsLibrary);
     for (int i = 0; i < warnings.length; i++) {
       warnings[i] = warnings[i].reassign(location);
     }
-    return new ArrayRope<>(warnings);
+    return warnings;
   }
 
   public static WithWarnings appendTo(Object target, ArrayRope<Warning> warnings) {
     if (target instanceof WithWarnings) {
+      return ((WithWarnings) target).append(warnings.toArray(Warning[]::new));
+    } else {
+      return new WithWarnings(target, warnings.toArray(Warning[]::new));
+    }
+  }
+
+  public static WithWarnings appendTo(Object target, Warning[] warnings) {
+    if (target instanceof WithWarnings) {
       return ((WithWarnings) target).append(warnings);
-    } else {
-      return new WithWarnings(target, warnings);
-    }
-  }
-
-  public static WithWarnings prependTo(Object target, ArrayRope<Warning> warnings) {
-    if (target instanceof WithWarnings) {
-      return ((WithWarnings) target).prepend(warnings);
-    } else {
-      return new WithWarnings(target, warnings);
-    }
-  }
-
-  public static WithWarnings prependTo(Object target, Warning[] warnings) {
-    if (target instanceof WithWarnings) {
-      return ((WithWarnings) target).prepend(warnings);
     } else {
       return new WithWarnings(target, warnings);
     }
@@ -133,9 +120,9 @@ public final class WithWarnings implements TruffleObject {
   Warning[] getWarnings(
       Node location, @CachedLibrary(limit = "3") WarningsLibrary warningsLibrary) {
     if (location != null) {
-      return getReassignedWarnings(location, warningsLibrary).toArray(Warning[]::new);
+      return getReassignedWarnings(location, warningsLibrary);
     } else {
-      return warnings.toArray(Warning[]::new);
+      return Warning.fromSetToArray(warnings);
     }
   }
 
@@ -152,6 +139,37 @@ public final class WithWarnings implements TruffleObject {
   @ExportMessage
   boolean hasSpecialDispatch() {
     return true;
+  }
+
+  public static class WarningEquivalence extends Equivalence {
+
+    @Override
+    public boolean equals(Object a, Object b) {
+      if (a instanceof Warning thisObj && b instanceof Warning thatObj) {
+        return thisObj.getSequenceId() == thatObj.getSequenceId();
+      }
+      return false;
+    }
+
+    @Override
+    public int hashCode(Object o) {
+      return (int)((Warning)o).getSequenceId();
+    }
+  }
+
+  @CompilerDirectives.TruffleBoundary
+  private EconomicSet<Warning> createSetFromArray(Warning[] entries) {
+    EconomicSet<Warning> set = EconomicSet.create(new WarningEquivalence());
+    set.addAll(Arrays.stream(entries).iterator());
+    return set;
+  }
+
+  @CompilerDirectives.TruffleBoundary
+  private EconomicSet<Warning> cloneSetAndAppend(EconomicSet<Warning> initial, Warning[] entries) {
+    EconomicSet<Warning> set = EconomicSet.create(new WarningEquivalence());
+    set.addAll(initial.iterator());
+    set.addAll(Arrays.stream(entries).iterator());
+    return set;
   }
 
   @Override