+++ engine/runtime/src/test/java/org/enso/interpreter/test/WarningsTest.java
@@ -1,17 +1,42 @@
 package org.enso.interpreter.test;
 
+import org.enso.interpreter.runtime.EnsoContext;
 import org.enso.interpreter.runtime.error.Warning;
 import org.enso.interpreter.runtime.error.WarningsLibrary;
 import org.enso.interpreter.runtime.error.WithWarnings;
+import org.enso.polyglot.LanguageInfo;
+import org.enso.polyglot.MethodNames;
+import org.graalvm.polyglot.Context;
+import org.junit.AfterClass;
 import org.junit.Assert;
-import static org.junit.Assert.assertEquals;
+import org.junit.BeforeClass;
 import org.junit.Test;
 
-public class WarningsTest {
+import static org.junit.Assert.assertEquals;
+
+public class WarningsTest extends TestBase {
+
+  private static Context ctx;
+
+  @BeforeClass
+  public static void initEnsoContext() {
+    ctx = createDefaultContext();
+  }
+
+  @AfterClass
+  public static void disposeContext() {
+    ctx.close();
+  }
+
   @Test
   public void doubleWithWarningsWrap() {
-    var warn1 = new Warning("w1", this, 1L);
-    var warn2 = new Warning("w2", this, 2L);
+    var ensoContext =
+        (EnsoContext)
+            ctx.getBindings(LanguageInfo.ID)
+                .invokeMember(MethodNames.TopScope.LEAK_CONTEXT)
+                .asHostObject();
+    var warn1 = Warning.create(ensoContext, "w1", this);
+    var warn2 = Warning.create(ensoContext, "w2", this);
     var value = 42;
 
     var with1 = WithWarnings.wrap(42, warn1);